LayoutTests/fast/html/details-no-summary-keyboard-show-hide-expected.txt

This test verifies that auto-generated <summary> in a <details> element is keyboard accessible.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS openAttribute("details") is false
Toggle <display> using Enter key:
PASS openAttribute("details") is true
PASS openAttribute("details") is false
Toggle <display> using Spacebar key:
PASS openAttribute("details") is true
PASS openAttribute("details") is false
PASS successfullyParsed is true

TEST COMPLETE
